+++ crates/ruma-events/src/beacon_info.rs
@@ -1,20 +1,22 @@
 //! Types for the `org.matrix.msc3489.beacon_info` state event, the unstable version of
 //! `m.beacon_info`.
+//!
+//! [MSC3489]: https://github.com/matrix-org/matrix-spec-proposals/pull/3489
 
-use std::time::Duration;
+use std::time::{Duration, SystemTime};
 
-use js_int::UInt;
 use ruma_common::{MilliSecondsSinceUnixEpoch, OwnedUserId};
 use ruma_macros::EventContent;
 use serde::{Deserialize, Serialize};
 
 use crate::location::AssetContent;
 
-/// `BeaconInfoEventContent` is a struct that represents the content of a beacon_info state
-/// event. It contains information about a live location share.
+/// The content of a beacon_info state
 #[derive(Clone, Debug, Deserialize, Serialize, EventContent)]
 #[cfg_attr(not(feature = "unstable-exhaustive-types"), non_exhaustive)]
-#[ruma_event(type = "org.matrix.msc3672.beacon_info", alias = "m.beacon_info", kind = State, state_key_type = OwnedUserId)]
+#[ruma_event(
+    type = "org.matrix.msc3672.beacon_info", alias = "m.beacon_info", kind = State, state_key_type = OwnedUserId
+)]
 pub struct BeaconInfoEventContent {
     /// The description of the location.
     ///
@@ -22,20 +24,20 @@ pub struct BeaconInfoEventContent {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub description: Option<String>,
 
-    /// `live` is a boolean that should be true when a user starts sharing location.
+    /// Whether the user starts sharing their location.
     pub live: bool,
 
-    /// The time when location sharing started, in milliseconds since the epoch.
+    /// The time when location sharing started.
     #[serde(rename = "org.matrix.msc3488.ts")]
     pub ts: MilliSecondsSinceUnixEpoch,
 
-    /// `timeout` represents the length of time in milliseconds that the location
-    /// will be live. So the location will stop being shared at `m.ts + timeout` milliseconds
-    /// since the epoch.
+    /// The duration that the location sharing will be live.
+    ///
+    /// Meaning that the location will stop being shared at `ts + timeout`.
     #[serde(default, with = "ruma_common::serde::duration::ms")]
     pub timeout: Duration,
 
-    /// `asset` is an [`AssetContent`] that this message refers to.
+    /// The asset that this message refers to.
     #[serde(default, rename = "org.matrix.msc3488.asset")]
     pub asset: AssetContent,
 }
@@ -57,25 +59,24 @@ impl BeaconInfoEventContent {
         }
     }
 
-    /// starts the beacon being live.
+    /// Starts the beacon_info being live.
     pub fn start(&mut self) {
         self.live = true;
     }
 
-    /// Stops the beacon from being live.
+    /// Stops the beacon_info from being live.
     pub fn stop(&mut self) {
         self.live = false;
     }
 
-    /// Checks if the beacon is currently live.
-    ///
-    /// This method calculates the current time and compares it with the beacon's start time plus
-    /// its timeout. If the beacon is not live or the current time is greater than the beacon's
-    /// start time plus its timeout, it returns false, indicating that the beacon is not live.
-    /// Otherwise, it returns true.
+    /// Start time plus its timeout, it returns `false`, indicating that the beacon is not live.
+    /// Otherwise, it returns `true`.
     pub fn is_live(&self) -> bool {
         self.live
-            && self.ts.get() + UInt::try_from(self.timeout.as_millis()).unwrap()
-                > MilliSecondsSinceUnixEpoch::now().get()
+            && self
+                .ts
+                .to_system_time()
+                .and_then(|t| t.checked_add(self.timeout))
+                .is_some_and(|t| t > SystemTime::now())
     }
 }