+++ crates/ruma-events/src/enums.rs
@@ -88,7 +88,7 @@ event_enum! {
         #[cfg(feature = "unstable-msc3381")]
         #[ruma_enum(ident = UnstablePollEnd)]
         "org.matrix.msc3381.poll.end" => super::poll::unstable_end,
-        #[cfg(all(feature = "unstable-msc3489", feature = "unstable-msc3488"))]
+        #[cfg(feature = "unstable-msc3489")]
         #[ruma_enum(alias = "m.beacon")]
         "org.matrix.msc3672.beacon" => super::beacon,
         "m.reaction" => super::reaction,
@@ -130,9 +130,9 @@ event_enum! {
         "m.room.topic" => super::room::topic,
         "m.space.child" => super::space::child,
         "m.space.parent" => super::space::parent,
-        #[cfg(all(feature = "unstable-msc3489", feature = "unstable-msc3488"))]
+        #[cfg(feature = "unstable-msc3489")]
         #[ruma_enum(alias = "m.beacon_info")]
-        "org.matrix.msc3672.beacon_info" => super::room::beacon_info,
+        "org.matrix.msc3672.beacon_info" => super::beacon_info,
         #[cfg(feature = "unstable-msc3401")]
         #[ruma_enum(alias = "m.call.member")]
         "org.matrix.msc3401.call.member" => super::call::member,
@@ -313,7 +313,7 @@ impl AnyMessageLikeEventContent {
     /// This is a helper function intended for encryption. There should not be a reason to access
     /// `m.relates_to` without first destructuring an `AnyMessageLikeEventContent` otherwise.
     pub fn relation(&self) -> Option<encrypted::Relation> {
-        #[cfg(all(feature = "unstable-msc3489", feature = "unstable-msc3488"))]
+        #[cfg(feature = "unstable-msc3489")]
         use super::beacon::BeaconEventContent;
         use super::key::verification::{
             accept::KeyVerificationAcceptEventContent, cancel::KeyVerificationCancelEventContent,
@@ -367,7 +367,7 @@ impl AnyMessageLikeEventContent {
             | Self::UnstablePollEnd(UnstablePollEndEventContent { relates_to, .. }) => {
                 Some(encrypted::Relation::Reference(relates_to.clone()))
             }
-            #[cfg(all(feature = "unstable-msc3489", feature = "unstable-msc3488"))]
+            #[cfg(feature = "unstable-msc3489")]
             Self::Beacon(BeaconEventContent { relates_to, .. }) => {
                 Some(encrypted::Relation::Reference(relates_to.clone()))
             }